if donations > 25% of target user can't delete the project

+++ projects/views.py
@@ -1,8 +1,9 @@
 from rest_framework import status
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
-from projects.helpers import validate_image_extension
+from django.db.models import Sum
 
+from projects.helpers import validate_image_extension
 from projects.models import Project , Image
 from projects.serializers import ProjectSerializer, DetailedProjectSerializer
 
@@ -46,6 +47,8 @@ def api_update_project(request, id):
 @api_view(['DELETE'])
 def api_delete_project(request, id):
     project = Project.objects.get(id=id)
+    if project.donation_set.aggregate(Sum('amount'))['amount__sum'] > project.total_target * 0.25:
+        return Response(status=status.HTTP_400_BAD_REQUEST)
     project.delete()
     return Response('Project deleted', status=status.HTTP_200_OK)
